put refactored code in CamusSweeper.java
